requirements.txt
Flask-SQLAlchemy==2.4.4
Flask==1.1.2
Flask-WTF==0.14.3
flask-executor==0.9.3
flask-cors==3.0.9
apispec==4.0.0
apispec-webframeworks==0.5.2
gdal==3.1.4
pyproj==2.6.1
psycopg2==2.8.4
Pillow==7.1.2
PyWavelets==1.1.1
qrcode==6.1
geovaex==0.2.0
pyzbar==0.1.8
